pyproject.toml
[project]
name = "social-media-agent"
version = "0.1.0"
description = "Social media agent for Notion, OpenRouter, and Mastodon integration"
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "notion-client>=2.2.1",
    "openai>=1.12.0",
    "Mastodon.py>=1.8.1",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "replicate>=0.25.0",
    "requests>=2.31.0",
    "pillow>=10.0.0",
    "python-telegram-bot>=21.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "apscheduler>=3.10.4",
    "pydantic-settings>=2.1.0",
    "croniter>=2.0.0",
    "fastembed>=0.7.4",
    "sqlite-vec>=0.1.6",
    "tiktoken>=0.12.0",
]

[project.scripts]
social-media-agent = "src.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
